# Conflicts:
#	build.gradle
#	src/integrationTest/java/uk/gov/hmcts/reform/jps/repository/SittingRecordRepositoryTest.java
#	src/integrationTest/java/uk/gov/hmcts/reform/jps/services/SittingRecoredServiceITest.java
#	src/integrationTest/resources/sql/add_sitting_record_status_history.sql
#	src/main/java/uk/gov/hmcts/reform/jps/repository/SittingRecordRepository.java
#	src/main/java/uk/gov/hmcts/reform/jps/repository/SittingRecordRepositorySearchImpl.java
#	src/main/java/uk/gov/hmcts/reform/jps/services/SittingRecordService.java
#	src/main/java/uk/gov/hmcts/reform/jps/services/StatusHistoryService.java
#	src/test/java/uk/gov/hmcts/reform/jps/controllers/SittingRecordControllerTest.java
#	src/test/java/uk/gov/hmcts/reform/jps/repository/SittingRecordRepositorySearchImplTest.java
#	src/test/java/uk/gov/hmcts/reform/jps/services/SittingRecordServiceTest.java

+++ src/integrationTest/java/uk/gov/hmcts/reform/jps/repository/StatusHistoryRepositoryTest.java
@@ -7,8 +7,10 @@ import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabas
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.data.domain.Sort;
 import org.springframework.test.context.ActiveProfiles;
+import uk.gov.hmcts.reform.jps.AbstractTest;
 import uk.gov.hmcts.reform.jps.domain.SittingRecord;
 import uk.gov.hmcts.reform.jps.domain.StatusHistory;
+import uk.gov.hmcts.reform.jps.model.JpsRole;
 import uk.gov.hmcts.reform.jps.model.StatusId;
 
 import java.time.LocalDate;
@@ -19,67 +21,58 @@ import java.util.List;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static uk.gov.hmcts.reform.jps.model.StatusId.RECORDED;
 
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
 @DataJpaTest
 @ActiveProfiles("itest")
-class StatusHistoryRepositoryTest {
+class StatusHistoryRepositoryTest extends AbstractTest {
 
     @Autowired
     private StatusHistoryRepository historyRepository;
     @Autowired
     private SittingRecordRepository recordRepository;
 
-    private StatusHistory persistedStatusHistory;
-
-    private StatusHistory statusHistory;
+    private SittingRecord persistedSittingRecord;
+    private StatusHistory persistedStatusHistoryRecorded;
+    private StatusHistory statusHistoryRecorded;
+    private StatusHistory statusHistorySubmitted;
 
     @BeforeEach
     public void setUp() {
-        SittingRecord sittingRecord = SittingRecord.builder()
-            .sittingDate(LocalDate.now().minusDays(2))
-            .statusId(RECORDED)
-            .regionId("1")
-            .epimmsId("123")
-            .hmctsServiceId("ssc_id")
-            .personalCode("001")
-            .contractTypeId(2L)
-            .am(true)
-            .judgeRoleTypeId("HighCourt")
-            .createdDateTime(LocalDateTime.now())
-            .createdByUserId("jp-recorder")
-            .build();
-
-        statusHistory = StatusHistory.builder()
-            .statusId(RECORDED)
-            .changeDateTime(LocalDateTime.now())
-            .changeByUserId("jp-recorder")
-            .changeByName("John Doe")
-            .build();
-
-        sittingRecord.addStatusHistory(statusHistory);
-        SittingRecord persistedSittingRecord = recordRepository.save(sittingRecord);
-        persistedStatusHistory = persistedSittingRecord.getStatusHistories().get(0);
+        SittingRecord sittingRecord = createSittingRecord(LocalDate.now().minusDays(2));
+        statusHistoryRecorded = createStatusHistory(sittingRecord.getStatusId(),
+                                                    JpsRole.ROLE_RECORDER.name(),
+                                                    "John Doe",
+                                                    sittingRecord);
+        sittingRecord.addStatusHistory(statusHistoryRecorded);
+        persistedSittingRecord = recordRepository.save(sittingRecord);
+        persistedStatusHistoryRecorded = persistedSittingRecord.getStatusHistories().get(0);
     }
 
     @Test
     void shouldSaveStatusHistory() {
-        assertThat(persistedStatusHistory).isNotNull();
-        assertThat(persistedStatusHistory.getId()).isNotNull();
-        assertThat(persistedStatusHistory).isEqualTo(statusHistory);
+
+        assertThat(persistedStatusHistoryRecorded).isNotNull();
+        assertThat(persistedStatusHistoryRecorded.getId()).isNotNull();
+        assertThat(persistedStatusHistoryRecorded).isEqualTo(statusHistoryRecorded);
     }
 
     @Test
     void shouldUpdateStatusHistoryWhenRecordIsPresent() {
 
         Optional<StatusHistory> optionalSettingHistoryToUpdate = historyRepository
-            .findById(persistedStatusHistory.getId());
+            .findById(persistedStatusHistoryRecorded.getId());
         assertThat(optionalSettingHistoryToUpdate).isPresent();
 
-        StatusHistory settingHistoryToUpdate = optionalSettingHistoryToUpdate.get();
-        settingHistoryToUpdate.setChangeDateTime(LocalDateTime.now());
-        settingHistoryToUpdate.setChangeByUserId("jp-submitter");
+        StatusHistory settingHistoryToUpdate = null;
+        if (optionalSettingHistoryToUpdate.isPresent()) {
+            settingHistoryToUpdate = optionalSettingHistoryToUpdate.get();
+            settingHistoryToUpdate.setChangeDateTime(LocalDateTime.now());
+            settingHistoryToUpdate.setChangeByUserId(JpsRole.ROLE_SUBMITTER.name());
+        }
 
         StatusHistory updatedStatusHistory = historyRepository.save(settingHistoryToUpdate);
         assertThat(updatedStatusHistory).isNotNull();
@@ -87,7 +80,7 @@ class StatusHistoryRepositoryTest {
     }
 
     @Test
-    void shouldReturnEmptyWhenHistorydNotFound() {
+    void shouldReturnEmptyWhenHistoryNotFound() {
         Optional<StatusHistory> optionalSettingHistoryToUpdate = historyRepository.findById(100L);
         assertThat(optionalSettingHistoryToUpdate).isEmpty();
     }
@@ -95,15 +88,68 @@ class StatusHistoryRepositoryTest {
     @Test
     void shouldDeleteSelectedHistory() {
 
-        Optional<StatusHistory> optionalSettingHistoryToUpdate = historyRepository
-            .findById(persistedStatusHistory.getId());
+        Optional<StatusHistory> optionalSettingHistoryToUpdate =
+            historyRepository.findById(persistedStatusHistoryRecorded.getId());
         assertThat(optionalSettingHistoryToUpdate).isPresent();
 
-        StatusHistory settingHistoryToDelete = optionalSettingHistoryToUpdate.get();
-        historyRepository.deleteById(settingHistoryToDelete.getId());
+        StatusHistory settingHistoryToDelete = null;
+        if (optionalSettingHistoryToUpdate.isPresent()) {
+            settingHistoryToDelete = optionalSettingHistoryToUpdate.get();
+            if (null != settingHistoryToDelete && null != settingHistoryToDelete.getId()) {
+                historyRepository.deleteById(settingHistoryToDelete.getId());
+            }
+        }
 
         optionalSettingHistoryToUpdate = historyRepository.findById(settingHistoryToDelete.getId());
         assertThat(optionalSettingHistoryToUpdate).isEmpty();
+    }
+
+    @Test
+    void shouldFindCreatedStatus() {
+
+        statusHistorySubmitted = createStatusHistory(StatusId.SUBMITTED,
+            JpsRole.ROLE_RECORDER.name(),
+            "Matthew Doe",
+            persistedSittingRecord);
+        persistedSittingRecord.addStatusHistory(statusHistorySubmitted);
+        historyRepository.save(statusHistorySubmitted);
+        persistedSittingRecord = recordRepository.save(persistedSittingRecord);
+
+
+        StatusHistory statusHistoryPublished = createStatusHistory(StatusId.PUBLISHED,
+                                                     JpsRole.ROLE_RECORDER.name(),
+                                                     "Matthew Doe",
+                                                     persistedSittingRecord);
+        persistedSittingRecord.addStatusHistory(statusHistoryPublished);
+        historyRepository.save(statusHistoryPublished);
+        persistedSittingRecord = recordRepository.save(persistedSittingRecord);
+
+        persistedStatusHistoryRecorded = persistedSittingRecord.getFirstStatusHistory();
+
+        StatusHistory statusHistoryFound = historyRepository
+            .findStatusHistoryAsc(persistedSittingRecord.getId()).get(0);
+        assertNotNull(statusHistoryFound, "Could not find any status history.");
+        assertEquals(statusHistoryFound, statusHistoryRecorded, "Not the expected " + StatusId.SUBMITTED
+            + " status history!");
+    }
+
+    @Test
+    void shouldFindLatestStatus() {
+
+        statusHistorySubmitted = createStatusHistory(StatusId.SUBMITTED,
+                                                     JpsRole.ROLE_RECORDER.name(),
+                                                     "Matthew Doe",
+                                                     persistedSittingRecord);
+        persistedSittingRecord.addStatusHistory(statusHistorySubmitted);
+        historyRepository.save(statusHistorySubmitted);
+        persistedSittingRecord = recordRepository.save(persistedSittingRecord);
+        persistedStatusHistoryRecorded = persistedSittingRecord.getFirstStatusHistory();
+
+        StatusHistory statusHistoryFound = historyRepository
+            .findStatusHistoryDesc(persistedSittingRecord.getId()).get(0);
+        assertNotNull(statusHistoryFound, "Could not find any status history.");
+        assertEquals(statusHistoryFound, statusHistorySubmitted, "Not the expected " + StatusId.SUBMITTED
+            + " status history!");
     }
 
     @Test
@@ -142,8 +188,6 @@ class StatusHistoryRepositoryTest {
             .contractTypeId(2L)
             .am(true)
             .judgeRoleTypeId("HighCourt")
-            .createdDateTime(LocalDateTime.now())
-            .createdByUserId("jp-recorder")
             .build();
     }
